removes old profile/ directory.
