+++ src/account/profile/Profile.jsx
@@ -139,3 +139,22 @@ export const Profile = () => {
 };
 
 export default Profile;
+
+/* import React from "react";
+import {
+  useAccountContext,
+  useAccountUpdateContext,
+} from "../contexts/AccountContext";
+
+export default function Profile() {
+  const { userType } = useAccountContext();
+  const { updateUserType } = useAccountUpdateContext();
+
+  return (
+    <div>
+      <div>Profile : {userType}</div>
+      <button onClick={updateUserType}>Click to change user type</button>
+    </div>
+  );
+}
+*/
